Report current tagged version when no commit is available
If the current elvish does not report a commit in its version info (e.g. 0.11-322-gf82458df vs 0.12-rc1), then the base tag is returned by `current-commit` (which has been renamed to `current-commit-or-tag`). This string can also be used in the commit queries to GitHub, so everything else still works the same.