Fixed "tag only" case (#4)
* Fixed "tag only" case

Fixed the case when there are no commits since the last tag (i.e. when total-commits is 0), to produce the same version as the Elvish Makefile does with "git describe --tags --always --dirty=-dirty".

* Report current tagged version when no commit is available

If the current elvish does not report a commit in its version info (e.g. 0.11-322-gf82458df vs 0.12-rc1), then the base tag is returned by `current-commit` (which has been renamed to `current-commit-or-tag`). This string can also be used in the commit queries to GitHub, so everything else still works the same.